pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "mcp>=1.25.0",
    "python-dotenv>=1.2.1",
]
